refactor: consolidate service registrations
